Fix build on 32-bit Windows; consolidate CI platform checks
This commit fixes the build of the `wasmtime` crate and CLI on 32-bit
Windows. This required adding some extra bits for trap handling and a
stub for "unimplemented" for unwind information registration. This then
additionally refactored our CI configuration to place all of the
cross-compile checks in one matrix instead of spread out across a few
jobs.

Closes #9882

prtest:platform-checks
